(postgres-infra-2) Improve and Consolidate Postgres Test Infrastructure
Summary:
This ports the infra from the airline demo into a shared test
module and makes some improvements to it. The airline demo itself
will be moved to this in the next diff in this stack.

1) For local development this does not spin up or down containers
if the postgres container in question is already running. This improves
cycle time dramaticaly. If the container is spun up we do not spin it
down. This is a debatable decision as it leaves the developer's machine
polluted with a running container. However I think the tradeoff is worth
it.

2) There was a race condition in postgres startup hidden by the long
startup cost of airline demo (due to Spark). Namely the test fixture
proceeded the moment the container was running but did not ensure that
postgres initialization was complete. This fixes the issue by polling
the postgres container for a db connect until it is done successful.
Only then does the fixture allow the test case to proceed.

Depends on D975

Test Plan: buildkite

Reviewers: #ft, max

Reviewed By: #ft, max

Differential Revision: https://dagster.phacility.com/D976