(new-dict-api-1) Move current Dict to global object with __call__
Summary:
This is the beginning of a sequence of diffs that move to new dict api.

The first step is make the top level Dict function a callable object. This will make
it so that we can use both the callable syntax as well as the bracket operators.

Test Plan: BK

Reviewers: #ft, alangenfeld, sashank

Reviewed By: #ft, alangenfeld

Differential Revision: https://dagster.phacility.com/D1016