refactor compute log manager to pull out abstract interface
Summary:
In order to support both local FS and s3 implementations, we need to extract a
common interface.  This diff extracts the local FS implementation, and also yanks out
all the subscription / file watching management into a separate class.

The plan is to have the s3 implementation rely on the local FS for the subscriptions /
in-flight capture, but upload to s3 as soon as the compute is done.

Depends on D1035

Test Plan: Ran unittests

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D1038