Add a floating context panel to the Dagit config editor
Summary:
This diff adds a floating panel to the right side of the config editor in the Execute tab that shows the configType schema at the insertion point. This leverages the existing <ConfigTypeSchema/> component that powers the Type Explorer and uses the same data we pull for autocompletion.

I also memoized a few components that were updating on every keystroke (including a dropdown, which resizes itself and leads to double renders). This approximately doubles typing performance in the config editor when there's a good chunk of text there. Almost all of the remaining time (still borderline unacceptable) is spent in Codemirror, so it's less clear what we can do about that.

Test Plan: Run tests

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D1057