(1/n) Create scheduler decorator and SchedulerBuilder
Summary:
Introducing an `@scheduler` decorator for defining a `Scheduler`.

Previously, the user defined a function that took an `artifacts_dir` argument, and returned a `Scheduler` which used the argument. This was done because `artifacts_dir` could only be resolved from the `Instance` at runtime, but the user-facing syntax made no sense.

Instead, the user uses the `@scheduler` decorator on a function that returns a list of `SchedulerDefinition`s, and we take care of passing the `artifacts_dir` to the `Scheduler` internally.

Also introducing a `SchedulerBuilder` here, which will soon be used to store the `ScheduleDefinition`s and help clean up the `Scheduler` interface.

Test Plan: unit

Reviewers: #ft, schrockn

Reviewed By: #ft, schrockn

Subscribers: schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D1105