(add-cancellation-4) Dramtically simplify SubprocessExecutionManager
Summary:
By levaraging the new instance infrastructure we can
substantially simplify the implementation of the
SubprocessExecutionManager. The original version of the class was
written before external run and event storage, and therefore required a
fairly contorted interaction using multiprocessing in order to
understand what was happening in subprocesses. This is not longer
required.

The new strategy to launch processes and let their logging into the
instance take care of mostly everything. You'll notice that instead of
pumping messages across the multiprocessing queue we are just inserting
them into the instance in the worker process directly.

The one complication here is to detect the case where a process hosting
a pipeline run has crashed unexpectedly leaving the run in a partially
completed state. The solution here is maintain an dictionary that maps
run id to the process associated with it. We then poll this to see
if a process has died. If the process is dead with the run in a
non-terminal state, we assume the process has crashed and report that to
user (and also mark the pipeline as failed).

Let's pour one out on the curb for the reverse semphamore after this is
merged.

Test Plan:
BK, Use dagit. Invoke many copies of the sleepy pipeline an
log spew pipeline and see expected behavior. Also invoke crashy_pipeline
and ensure the crash message appears in dagit.

Reviewers: #ft, alangenfeld, max

Reviewed By: #ft, max

Subscribers: prha

Differential Revision: https://dagster.phacility.com/D1098