Add limit to SchedulesRootQuery
Summary:
The `SchedulesRootQuery` previously queried for all runs that corresponded to a schedule.  In dagit, however, we only display the ten most recent runs and have a `+ XX more` link that navigates to the runs page with the schedule tag filter. When a schedule accumlates a large number of runs, the page takes too long to load.

This diff adds a limit to the query so that we only query for the runs we want to display on the scheduler page, and adds a new `runsCount` field to get the total number of runs for a schedule.

Test Plan: Load dagit, get same results as before

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: natekupp, alangenfeld

Differential Revision: https://dagster.phacility.com/D1276