Loosen future dependency in outside of dagster-airflow
Summary:
I don't believe we need to account for airflow's depedency
constraints in core dagster. If we loosen the dep here (I'm not sure
exactly what we rely on in dagster core) then when dagster-airflow
installs that will deal with their weird issues.

This is a mitigation. We should really be executing execution steps
in the airflow integration in a totally different virtual environment,
so we don't repeat Airflow's mistake of intermixing the orchestration
and compute environments.

Resolves https://github.com/dagster-io/dagster/issues/1869

Test Plan: BK

Reviewers: #ft, max, natekupp

Reviewed By: #ft, natekupp

Differential Revision: https://dagster.phacility.com/D1324