Update get_runs_with_matching_tag method on RunStorage to query for multiple tags
Summary:
Previously, `get_runs_with_matching_tag` took args `key` and `value`, which lets you query for runs by one tag. This diff updates the method to take in a list of `(key, value)` tuples to query the `RunStorage` for runs that have all of the given tags.

This is a breaking change to `RunStorage`, but this is also the first diff in a stack to consolidate these various `RunStorage` methods into a single method. In the next release, we can consolidate all these breaking changes into one change that introduces the new method.

Test Plan: unit

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: alangenfeld, themissinghlink

Differential Revision: https://dagster.phacility.com/D1332