[dagster-airflow] Fix k8s pod operator in dagster-airflow and get it under test
Summary:
well, it only took me 30 diffs, but this is now working and ready for review.

This diff fixes the Dagster k8s operator and provides a `kind`-based unit test, getting it under test.

To speed up this test, I've also moved the Docker image build step out to the nightly; it was getting extremely slow.

Test Plan: unit

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D1429