Kill is_secret
Summary:
This currently doesn't do anything. I think it is better not to have it
rather than have it and have it deliver no value to users. It will easy enough
to readd this part of it later. In order to do this right we have to do a non-trivial
change to the logging system and the config editor, and it's not really a priority
at the moment.

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1441